% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seminar_script.R
\name{gene_table}
\alias{gene_table}
\title{import count and sample file}
\usage{
gene_table(count_file, sample_file)
}
\arguments{
\item{count_file}{is the count file for RNA-seq experiment}

\item{sample_file}{is the sample file where you define which group each sample belongs to. assumes the file has header and is separated with tab}
}
\value{
count and sample file  as data tables assumes the file has header and is separated with tab
}
\description{
import count and sample file
}
\examples{
gene_table("E-MTAB-2523.counts.txt", "E-MTAB-2523_sample table.txt")
}
