% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seminar_script_enrichmentAnlasys.R
\name{GO_pathway}
\alias{GO_pathway}
\title{Gene Ontology GO enrichment analysis}
\usage{
GO_pathway(gene_type = "SYMBOL")
}
\arguments{
\item{gene_type}{tells the function what kind Gene ID it is, take gene symbol or ENTREZ ID}
}
\value{
a table with over represented gene
}
\description{
Gene Ontology GO enrichment analysis
}
\examples{
GOpathway(gene_type = "SYMBOL")
GOpathway(gene_type = "ENTREZID")
}
