% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seminar_script_enrichmentAnlasys.R
\name{KEGG_pathway}
\alias{KEGG_pathway}
\title{KEGG pathway enrichment analysis}
\usage{
KEGG_pathway(gene_type = "SYMBOL")
}
\arguments{
\item{gene_type}{tells the function what kind Gene ID it is, take gene symbol or ENTREZ ID}
}
\value{
data table that shows enrich pathway in a genne set
}
\description{
KEGG pathway enrichment analysis
}
\examples{
KEGG_pathway(gene_type = "SYMBOL")
KEGG_pathway(gene_type = "ENTREZID")
}
