% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seminar_script.R
\name{low_gene_filtering}
\alias{low_gene_filtering}
\title{filter out the low count of the count table}
\usage{
low_gene_filtering(cutoff = 1, count_table, sample_table)
}
\arguments{
\item{cutoff}{tells the function to filter out all below the cutoff}

\item{count_table}{is the count file for RNA-seq experiment. assumes the file has header and is separated with tab}

\item{sample_table}{is the sample file where you define which group each sample belongs to.assumes the file has header and is separated with tab}
}
\value{
filtered counts
}
\description{
filter out the low count of the count table
}
\examples{
gene_table(cutoff = 1,"E-MTAB-2523.counts.txt", "E-MTAB-2523_sample table.txt")
}
