% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seminar_script_enrichmentAnlasys.R
\name{GO_pathway}
\alias{GO_pathway}
\title{Gene Ontology GO enrichment analysis}
\usage{
GO_pathway(count_table, sample_table, gene_type = "SYMBOL")
}
\arguments{
\item{gene_type}{tells the function what kind Gene ID it is, take gene symbol or ENTREZ ID}

\item{count_file}{is the count file for RNA-seq experiment, assumes the file has header and is separated with tab}

\item{sample_file}{is the sample file where you define which group each sample belongs to. assumes the file has header and is separated with tab}
}
\value{
Export the results of the Gene Ontology enrichment analysis analysis to a CVS file
}
\description{
Gene Ontology GO enrichment analysis
}
\examples{
GOpathway(gene_type = "SYMBOL")
GOpathway(gene_type = "ENTREZID")
}
